Join message now edits existing message if present
If multiple users join with no other activity on the channel then the
join messages will be combined into a single message for brevity.

TODO: it says the command has failed, but it works correctly and no
error is logged in the console. This may be a consequence of how I'm
testing it.